requirements.txt
langchain
langchain_openai
langchain_community
openai
pydantic
fastapi
uvicorn
psycopg2-binary
streamlit
requests
python-dotenv
pytest
pytest-asyncio
httpx
pytest-mock
pytest-cov
faker
factory-boy
testcontainers
sqlalchemy-utils